+++ apps/web/app/(use-page-wrapper)/(main-nav)/availability/loading.tsx
@@ -1,4 +1,4 @@
-import AvailabilityLoader from "app/(use-page-wrapper)/(main-nav)/availability/clientLoader";
+import AvailabilityLoader from "app/(use-page-wrapper)/(main-nav)/availability/skeleton";
 
 export default function Loading() {
   return <AvailabilityLoader />;
